texel: Improved compatibility with visual studio.
git-svn-id: file:///home/petero/vc/svnrepos/texel/trunk@486 4abc0a36-5a1e-4bec-8e7d-414f7125f69d

+++ texellib/src/transpositionTable.hpp
@@ -298,7 +298,7 @@ TranspositionTable::TTEntry::setBits(int first, int size, unsigned int value) {
 inline unsigned int
 TranspositionTable::TTEntry::getBits(int first, int size) const {
     U64 sizeMask = ((1ULL << size) - 1);
-    return (data >> first) & sizeMask;
+    return (unsigned int)((data >> first) & sizeMask);
 }
